--- src/main/java/FridayException.java
@@ -1,3 +1,0 @@
-public class FridayException extends Exception {
-    FridayException(String msg) { super(msg); }
-}
