+++ src/main/java/friday/exception/FridayException.java
@@ -1,0 +1,7 @@
+package friday.exception;
+
+public class FridayException extends Exception {
+    public FridayException(String msg) {
+        super(msg);
+    }
+}
