+++ src/main/java/friday/model/ToDo.java
@@ -1,0 +1,10 @@
+package friday.model;
+
+public class ToDo extends Task {
+    public ToDo(String d) {
+        super(d);
+    }
+    String typeIcon() {
+        return "[T]";
+    }
+}
