+++ src/main/java/friday/storage/Storage.java
@@ -1,3 +1,7 @@
+package friday.storage;
+
+import friday.model.Task;
+
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.*;
