+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello! I'm Friday
+ Hello! I'm friday.app.Friday
  What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
